chore: move to t.ts files to address skipLibCheck issue
