+++ src/triggers/click.ts
@@ -1,4 +1,4 @@
-import type { NavigationTrigger } from '../types.js';
+import type { NavigationTrigger } from '../types.t.js';
 import { fireRouterEvent } from '../utils.js';
 
 /* istanbul ignore next: coverage is calculated in Chrome, this code is for IE */
